SurfaceFlinger: Set the buffer size in the GraphicBufferAlloc
When setting the buffer size, store the required size, and allocate
the buffer of the required size in GraphicBufferAlloc, instead of
SurfaceTexture. We do this because SurfaceTexture does not always
run in the server process space. GraphicBufferAlloc always runs in the
server process.

(cherry picked from commit 5221ca13baadba79d9092269313f3682304451d2)

Change-Id: I2dfc478486c9068d58716c32bc13a107d5a28720