frameworks/base: creating tombstones for list of processes on framework reboot
On the framework reboot due to watchdog killing system_server (one the locks
in the system_server thread was held for more than 60 sec) the watchdog thread
creates the tombstone file for system_server process. But many times the reason
for the framework reboot could be because of some other process going to bad
state and the system_server was unbale to proceed further.

This patch allows the user to provide the process names with "," separated list
in the property variable "sys.watchdog.extraprocnames" to generate the
tombstone files for these processes on the framework reboot.

(cherry picked from commit bc66849d53e388c0791216493e189c9ca1df0362)

Change-Id: Id1059a1508662ce6ac1e6e2150b4f268a9470859