libstagefright: Fix for auto resume issue when QCELP clips are paused
-Issue: The issue happens with a timing where the playback is paused
and there is a wait on the buffer filled. The wait timed out as the
playback is in pause state notifying TIMED OUT error and as a result
the playback is abruptly closed follwed by the next song playback
from the list.

-Test Scenario: Device switch from headet to speaker when playing
QCP clips and wait for 5 seconds. The issue is observed at random

-Fix: Handled the TIMEDOUT by ignoring the timed out error if the
playback is in pause state .

CRs-Fixed: 333397
(cherry picked from commit 661faf73b904c23d83f759088eacddd250e96576)

Change-Id: I3bd1b00b91c14f74967d7189ed43b0505f7ea0a5