SystemUI: Consider eri only when in service
Add in service check to isCdmaEri.
Telephony UI considers eri value
for roaming display only during in service.
Without this when not in service if RIL
sends prl loaded , telephony
would compute eri and show roaming
because the roaming related information
in registration state will not be valid
during out of service.

CRs-Fixed: 339297
Change-Id: I9380d19b7dc763ed43c8b8442150674e48d9d595