Wifi: Make updating V6 addresses asynchronous
Add a new command that will be handled in the Connected State
to update the link properties with the IPv6 addresses and
route.

When the link properties are to be updated, don't wait for the
IPv6 related information, instead post the above command to
be handled at a later time.

Also handle IPv6 renewal when lease time expires.

Change-Id: Ic1cb7d307d1f0bf5485c9a931784237995fa6232
CRs-Fixed: 349943
(cherry picked from commit 2e79416b936739e8af1a2534153869fb3d7bb8e6)