libstagefright: Fix memory leak in video playback start failure
Output buffers are not freed when video playback fails
when starting playback, causing a memory leak.

In the case of a codec initialization failure, we now make
sure that the omx component is properly transitioned to
loaded state and that the buffers are freed.

CRs-Fixed: 360964
(cherry picked from commit 37fa976da8bbeff3d5edf55785224b1c389154d5)

Change-Id: I9a0efe80564a3fca1847f5e5c08b671fa78927af